Revert confluence connector
